orquestrador/docker_requirements.txt
apache-airflow-providers-apache-spark
apache-airflow-providers-openlineage>=1.8.0
apache-airflow-providers-postgres
apache-airflow-providers-sqlite
